Delete redundant functions
Random sound is stored in a global variable. Therefore, all getRandomSound functions are not necessary. Also, user's answer is retrieved from button's id instead of text which allows divergence of id and text, which allows using '#' for sharp notes and '+' for representing them in the sound src and the button id.